Time: 1386 ms (22.96%), Space: 190 MB (15.26%) - LeetHub

+++ 2421-number-of-good-paths/2421-number-of-good-paths.cpp
@@ -60,7 +60,7 @@ public:
           d.merge(u,v);
         }
       }
-      map<int,ll> cnt;
+      map<int,int> cnt;
       for (auto u : a) {
         cnt[d.leader(u)]++;
       }
@@ -70,4 +70,4 @@ public:
     }
     return ans;
   }
-};
+};
